First test API

+++ retz-common/src/main/java/io/github/retz/protocol/data/Job.java
@@ -37,7 +37,7 @@ public class Job {
     private String finished;
     private Properties props;
     private int result = -1;
-    private int id;
+    private long id;
     private String url;
     private String reason;
     private int retry; // How many retry now we have
@@ -82,7 +82,7 @@ public class Job {
                @JsonProperty("finished") String finished,
                @JsonProperty("env") Properties props,
                @JsonProperty("result") int result,
-               @JsonProperty(value = "id", required = true) int id,
+               @JsonProperty(value = "id", required = true) long id,
                @JsonProperty("url") String url,
                @JsonProperty("reason") String reason,
                @JsonProperty("retry") int retry,
@@ -149,7 +149,7 @@ public class Job {
     }
 
     @JsonGetter("id")
-    public int id() {
+    public long id() {
         return id;
     }
